edit the code to fit the right requierment without mearg
